+++ app/src/main/java/woowacourse/shopping/presentation/home/adapter/LoadingViewHolder.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.presentation.home.adapter
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.databinding.ItemLoadMoreBinding
 import woowacourse.shopping.presentation.home.HomeActionHandler
-import woowacourse.shopping.presentation.home.LoadStatus
+import woowacourse.shopping.presentation.uistate.LoadStatus
 
 class LoadingViewHolder(
     private val binding: ItemLoadMoreBinding,
