+++ app/src/main/java/woowacourse/shopping/presentation/uistate/Order.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.presentation.cart
+package woowacourse.shopping.presentation.uistate
 
 import woowacourse.shopping.data.model.Product
