+++ app/src/main/res/layout/item_load_more.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="loadStatus"
-            type="woowacourse.shopping.presentation.home.LoadStatus" />
+            type="woowacourse.shopping.presentation.uistate.LoadStatus" />
 
         <variable
             name="homeActionHandler"
